Move updates workers to workers module.
