+++ spyder/utils/workers/__init__.py
@@ -5,3 +5,6 @@
 # Distributed under the terms of the MIT License
 # (see spyder/__init__.py for details)
 # -----------------------------------------------------------------------------
+
+from spyder.utils.workers.manager import (ProcessWorker, PythonWorker,
+                                          WorkerManager)
