+++ spyder/utils/workers/manager.py
