+++ spyder/utils/workers/updates.py
